Add files via upload
Here is the only file needed , and of course tesseract is needed to have the ocr attributes. In this case tesseract is in the "Program Files (x86)" folder but changing this specific line of the code already solves the problem.
To facilitate the usage of this app the ideal situation would be finding tesseract automatically but i could not solve this problem yet.
This code needs only to press one key to read the picture and then it turns itself off. I would love seeing an multi-thread system which could keep being on but only read when called. That's another problem i could not solve yet.